Remove high/low missile support
Remove all the horrible code to deal with having a single Lua controller deal with both missiles or torpedoes, which has never been tested. The much better way to have torpedo support would be with a separate Lua controller block and having this script restrict itself to given weapon groups, once a way to do that exists or can be found.

Also replaces half_pi with a tunable in the process of necessary renames.