Refactor; detect sphere of combat; chase unicorns
Break the loop out of SteerMissiles into Update so it can be shared with measurement mode.
Add target filtering for above/below water that was lost when high/low mode was removed. Reintroduces sea_crossover_tolerance.
Add an option to go for the best invalid target if there are no valid targets for a missile.
Add some missing doc.