Improvements to Heliairship control logic
Add new script for drive control, rather than using ACBs for it, which
is crude and wobbly. Requires hydroplanes to read control intent from
pilot/AI. Adds smoothing, and reacts to requests for hard turns with
pivoting in place without throwing the engines into reverse on one side
for minor forward course corrections.

Use same new logic to correctly detect orientation of propellers in the
balance script.

Add corporate overlord to copyright holders because that is the world we
live in.